requirements.txt
discord.py
flask
pymongo
python-dotenv